+++ android/src/main/java/company/tap/goSellSDKExamplee/GoSellSdKDelegate.java
@@ -25,6 +25,8 @@ import company.tap.gosellapi.GoSellSDK;
 import company.tap.gosellapi.internal.api.callbacks.GoSellError;
 import company.tap.gosellapi.internal.api.models.Authorize;
 import company.tap.gosellapi.internal.api.models.Charge;
+import company.tap.gosellapi.internal.api.models.Contract;
+import company.tap.gosellapi.internal.api.models.PaymentAgreement;
 import company.tap.gosellapi.internal.api.models.SaveCard;
 import company.tap.gosellapi.internal.api.models.Token;
 import company.tap.gosellapi.open.controllers.SDKSession;
@@ -579,6 +581,27 @@ public class GoSellSdKDelegate implements SessionDelegate {
         if (charge instanceof SaveCard) {
             System.out.println("Card  are Saved Succeeded : first six digits : " + ((SaveCard) charge).getCard().getFirstSix() + "  last four :" + ((SaveCard) charge).getCard().getLast4());
         }
+
+
+        if(((SaveCard) charge).getPaymentAgreement() != null) {
+            PaymentAgreement paymentAgreement = ((SaveCard) charge).getPaymentAgreement();
+            HashMap<String, Object> paymentAgreementMap = new HashMap<>();
+            paymentAgreementMap.put("id", paymentAgreement.getId());
+            paymentAgreementMap.put("type", paymentAgreement.getType());
+            paymentAgreementMap.put("total_payments_count", paymentAgreement.getTotalPaymentCount());
+            paymentAgreementMap.put("trace_id", paymentAgreement.getTraceId());
+            if(paymentAgreement.getContract() != null) {
+                Contract contract= paymentAgreement.getContract();
+                HashMap<String, Object> contractMap = new HashMap<>();
+                contractMap.put("id", contract.getId());
+                contractMap.put("customer_id", contract.getCustomerId());
+                contractMap.put("type", contract.getType());
+                paymentAgreementMap.put("contract",contractMap);
+            }
+            resultMap.put("payment_agreement", paymentAgreementMap);
+        }
+
+
         resultMap.put("card_first_six", ((SaveCard) charge).getCard().getFirstSix());
         resultMap.put("card_last_four",((SaveCard) charge).getCard().getLast4());
         resultMap.put("card_status",charge.getStatus().toString());